add clipping
