+++ gpx2db_pkg/gpx2db/sql/proximity-calc/0100_joinsegments_create_newpoints.sql
@@ -2,7 +2,14 @@
 --- by inserting points into segments and renumber the segment id
 
 insert into joined_points
-with base as (
+with clipped_points as (
+	SELECT
+		*
+	FROM track_points tp
+	where
+		{}
+),
+base as (
 select 
 	tp1.id,
     tp1.track_id,
@@ -18,17 +25,19 @@ select
     		or 
     		-- new segment starts, but only when distance bigger than threshold
     		( 
-                (tp1.track_segment_id != tp2.track_segment_id)
-    			and (ST_Distance(tp1.wkb_geometry::geography, tp2.wkb_geometry::geography) >= 30)
-            )
+          (tp1.track_segment_id != tp2.track_segment_id)
+    			and
+					(ST_Distance(tp1.wkb_geometry::geography, tp2.wkb_geometry::geography) >= 30)
+        )
     	then 1 --- marker 
     	else null 
     end as marker,
     tp1.wkb_geometry
-from track_points tp1 left join track_points tp2
+from clipped_points tp1 left join clipped_points tp2
 on
     tp1.id = tp2.id + 1 -- vergleiche mit vorhergehendem punkt
-where tp1.track_id = {}
+where
+	tp1.track_id = {}
 order by id
 )
 select 
@@ -41,3 +50,4 @@ select
 	end as segment_id,
     wkb_geometry
 from base;
+
